requirements.txt
feedparser==6.0.12
beautifulsoup4==4.14.2
requests==2.32.5
lxml==6.0.2
psycopg2-binary==2.9.9
google-generativeai==0.3.2
python-dotenv==1.0.0
